Fixed VMU read support. Added two utilities, vmu_dump and vmu_flash, to transfer a Flash image (a 128k .VMU file) from and to the VMU. Seems to work!
The implementation is now a little messy because the VMU can transmit up to 520 bytes to the Arduino in one response. I had assumed that the VMU would never send more than 256 bytes, so the code is based around single-byte counters. Unfortunately reads from VMU flash return 512 bytes of data (unlike writes, which must be performed 128 bytes at a time).

Rather than convert all the code to work with 16-bit lengths (a bit of a pain on an 8-bit microcontroller), I kept the PC->Arduino->VMU path 8-bit. The VMU->Arduino->PC path returns 16-bit lengths. This should be acceptable, as the maximum amount of data one can write to a VMU is 200 bytes (to display an image on the LCD).